services/web/pyproject.toml
[tool.poetry]
name = "trading-web"
version = "1.0.0"
description = "Algorithmic Trading Platform - Web Service"
authors = ["Trading Platform <dev@tradingplatform.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14"
django-environ = "^0.11"
django-cors-headers = "^4.3"
django-celery-beat = "^2.5"
channels = {extras = ["daphne"], version = "^4.0"}
channels-redis = "^4.1"
psycopg2-binary = "^2.9"
redis = "^5.0"
celery = {extras = ["redis"], version = "^5.3"}
pydantic = "^2.5"
pydantic-settings = "^2.1"
pandas = "^2.1"
polars = "^0.20"
numpy = "^1.26"
plotly = "^5.18"
scipy = "^1.12"
statsmodels = "^0.14"
requests = "^2.31"
websocket-client = "^1.7"
prometheus-client = "^0.19"
python-json-logger = "^2.0"
sentry-sdk = {extras = ["django"], version = "^1.39"}
gunicorn = "^21.2"
uvicorn = {extras = ["standard"], version = "^0.25"}
python-dateutil = "^2.8"
pytz = "^2024.1"
drf-spectacular = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
pytest-cov = "^4.1"
pytest-asyncio = "^0.23"
ruff = "^0.1"
black = "^23.12"
mypy = "^1.8"
django-stubs = "^4.2"
djangorestframework-stubs = "^3.14"
ipython = "^8.19"
pre-commit = "^3.6"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["tests"]
addopts = "-v --strict-markers --disable-warnings"

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*", "*/fixtures/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
